feat: Add, improve things, and finish simple DOS

+++ src/libs/math/pow.c
@@ -1,4 +1,4 @@
-long float pow(long number, long exp) {
+float pow(long number, long exp) {
     long result = number;
     while (exp > 1) {
         result = result * number;
