+++ src/libs/math/sqrt.c
@@ -1,6 +1,6 @@
-long float sqrt(long float number) {
-    long float result = number;
-    long float lastResult = 0;
+float sqrt(float number) {
+    float result = number;
+    float lastResult = 0;
     while (result != lastResult) {
         lastResult = result;
         result = (result + number / result) / 2;
